Builds a merge tool in Golang; validates everything works as expected
Instead of using a shell script for managing the merge cli tool, I decided to build a Golang app around it instead for ease of use, updating, management, and testing.

The project is functional and can be used as-is today, including uploading to R2. Ostensibly S3 uploads work too, but I haven't tried.